Console - Moved package .client.overview to more appropriate .client.dashboard
